+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 ### Changed
 - Use "omitempty" in marshaler settings with data definition structs.
 
+### Fixed
+- Fixed unrecognized Inventory system type in storage personalizer.
+
 ## [1.9.0] - 2024-02-04
 ### Added
 - New option "max_repeat_rolls" to set how many duplicate rows of rolled rewards can occur.
@@ -151,7 +154,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Add ChannelMessageAck message to proto definition.
 
 ### Fixed
-- Expose server functions for reward and roll in Hiro. 
+- Expose server functions for reward and roll in Hiro.
 
 ## [1.0.3] - 2023-08-10
 ### Added